requirements.txt
flask==2.2.3
boto3==1.26.99
python-dotenv==1.0.0